feat(dal): Add helper for checking if any of an AttributeResolver's siblings are set
This will be useful when we go to unset a particular AttributeResolver,
and want to know the answer to the question "am I the last
AttributeResolver with a value in my parent?". If the AttributeResolver
is the last child of the parent to have a value, then we'll also want to
unset the Parent (repeating this logic all the way to the root of the
lineage chain, or until we the answer to "was I the only child with a
value" is "no").