feat(dal,cyc,ver,lang): Pass ComponentView to veritech (recursive props)
Changes data passed to veritech to contain a recursive view of the props, instead of only the top layer that we depended on.

This implies in changing a lot of boilerplate from lang-js, cyclone, veritech and the dal. Including fixing tests that depended on the structure.

Also fixed a platform incompatibility (Mac's tempfolder is different than GNU/Linux ones).

For now we can't set a default value of a child prop without setting a default value for a parent prop. This should be fixed in a future task, so we just disabled default values for child props for now, as they weren't useful (they were just there to specify how to do it).