feat(web,sdf): Create List Secrets service
Fetches the secrets list when the Secret Panel is setup. This means it won't get updated, but to update we will need a rewrite, or a websocket message telling us a secret was created, and we can't have it without having the create secrets route.

Besides fetching the database for secrets (which we can't exactly test in the sdf layer yet), it also returns a mock secret to help testing things in the frontend, but it should be removed this week still.

We need tests for this service, but as we don't actually have a secret creation route it would be a really dumb test.

[sc-2280]