feat(dal/ver/cyc): Pass secrets in ComponentView as Sensitive data
Qualifications now receive parent/configuration components.

The parent components are parsed in FuncBinding and the secrets get extracted and unencrypted.
The secret is then passed as a SensitiveContainer<serde_json::Value> to cyclone and then lang-js.
Nats and lang-js are the only places where secrets could be exposed. We need to implement a filter in lang-js.
To deal with Nats we should encrypt the data e2e from dal to cyclone (symetric crypto should be ok).

Schemas/Components now get a ComponentKind, as credentials components need special treatment.

The biggest loose end this PR creates is that we can't actually know which property is a secret, as we need to unencrypt
the secrets the lowest in the stack possible to avoid leaking the data to the database, but by then we lose info about the prop.

So for now we are only decrypting the root props that are integers. Which sucks hard. But it's workable. This shouldn't be solved this week tho.

Example: if done in ComponentView::for_component_and_system we could check for prop.widget_kind() and only decrypt for WidgetKind::SecretSelect, but it would store the secret in the FuncBinding)

It's not clear if we want to remove the credential tag from the properties, as it's useful to the functions. For now we only remove on the code generation as it messes with the YAML. We need more thinking into this.