feat(dal): Auto-vivify, and -devivify parent AttributeResolvers as necessary when setting deeply nested Props
If we have a Schema like the following:

```
[
    { "name": String, "release_year": String },
]
```

When we attempt to add a new object to the array, we would like the
array to automatically exist if it doesn't already. Similarly, when we
unset `name` or `release_year` on an object in the array, that specific
object should become unset if and only if both `name` and `release_year`
are unset. This un-setting behavior should also continue up the tree to
the root if all objects in the array are unset, causing the array itself
to become unset.

This is preparing to migrate the behavior of
`Component.resolve_attribute` into two methods within
`AttributeResolver`.

`AttributeResolver::update_for_context` is intended for setting/updating
the appropriate `AttributeResolver` for a given context, and should be
the generally used method for setting/un-setting Props for a Component,
or Schema.

`AttributeResolver::insert_for_context` is intended for adding a new
element to a parent Array/Map, and uses
`AttributeResolver::update_for_context` under the hood to set the value
after it has created a new unset `AttributeResolver` at the appropriate
place.

[sc-2245]
[sc-2278]