feat(cyc,ver): Unify commands boilerplate
This removes a big part of the RPC boilerplate, as we unify the server
and the client part of calling a lang-js function into two generic structures
instead of having a copy for each command.

Now to create a new function kind you just need to create the
Request/ResultSuccess types, a function that calls Execute<Request, ResultSuccess>,
and configure veritech to allow enabling this new kind. (besides
exporting the Request/ResultSuccess/Component from cyclone to deadpool_cyclone
to veritech so the dal can access them)