feat(cyclone): Filter secrets in lang-js outputs
Collects secrets from a ComponentView by iterating over properties
of a ComponentKind::Credential and trying to identify DecryptedSecret
from it's fields. We only censor DecryptedSecret::secret_kind and
DecryptedSecret::message from ComponentViews.

Censored data is replaced for '[redacted]' and is matched by substring
so it could match with data out of context therefore exposing that we
have a secret just because some random substring happened to match it.

Filters lang-js stderr, function logs and iterates over the
success result and replaces the values. This last part might
not be wanted, but a qualification message might accidentally leak
a credential. So it seemed net positive to do so.

For now we are ignoring credential keys, but this might not be ideal.