feat(dal): Add support for Hashes to AttributePrototype
This renames `find_for_context` to `list_for_context`, as we won't
know if we should only ever be getting a single Prototype for a given
Context or not, without looking at the parent Prop of the prop that
we're interested in (Hashes and Arrays have the potential to have
multiple Prototypes for a single child Prop).

This also updates the query to get the Prototypes for a given Context
to include Key (the Hash key) in their uniqueness constraints, and to
remove the `LIMIT 1` to allow all appropriate results for the Context
to be returned.