chore: add configurable cyclone secret re-encryption to codebase
This is a rather large change, but ultimately its goal is to introduce
configurable cyclone re-encryption of secrets when SDF (our primary API
service) dispatches functions to be executed via Veritech (the function
execution consuming service), Cyclone (the function execution
executor), and ultimately a language server (currently in our case this
is only `lang-js` which executes the function written in a particular
language).

When secrets are added to our system, they are persisted via the DAL
(our data abstraction layer which manages persistence and related
business logic) as `EncryptedSecret`s. There is only one consumer of a
full `EncryptedSecret`, which is the function execution system which
requires these secrets to be decrypted so that they can be consumed and
used. Importantly, once a function execution is dispatched via a
Veritech client, there is no longer any system access to the database or
any other DAL behavior, meaning that everything required for the
function to execute must ship with the request.

In order to prevent decrypted secrets being transmitted over message
queues (currently we are using NATS) we would desire these secrets to be
transmitted in a protected fashion. To this end, when we prepare a
function execution request after decrypting secrets, we are
re-encrypting these secrets using key pair encryption with the
sodiumoxide library. Note that this an unrelated mechanism and key pair
used to persist the original `EncryptedSecret` at rest in the database.

Encryption and Decryption Locations
-----------------------------------

This re-encryption means that now we have a new key pair to manage when
deploying, running, developing, and testing our full system. Luckily,
this initial approach (that is, for now one known key pair for all
requests) is reasonably straight forward:

* SDF is the only actor in our system which can decrypt the
  `EncryptedSecret` and so needs the public side of the key pair as an
  `EncryptionKey`.
* Cyclone is the only consumer of these requests which requires the
  secret side of the key pair for decrypting, hence it needs a
  `DecryptionKey`.

We've attempted to call this key pair by 2 distinct nouns (i.e.
`EncryptionKey` and `DecryptionKey`) as they are never required by any
one system to both be provided. In other words, *only* Cyclone needs
only the `DecryptionKey` and *only* SDF needs only the `EncryptionKey`.

Encryption and Decryption Key Generation
----------------------------------------

While both of these related keys is technically a key pair, we're
keeping the names somewhat split and as a result generating these keys
involves 2 CLI options. Currently generating these keys are the
responsibility of the `sdf` program by providing both the
`--generate-cyclone-public-key-path` and
`--generate-cyclone-secret-key-path` flags. For example:

```sh
sdf \
  --generate-cyclone-public-key-path /run/sdf/cyclone_encryption.key \
  --generate-cyclone-secret-key-path /run/cyclone/decryption.key
```

Note that both options are required, and `sdf` will fail with a message
if this is not the case (remember, they are a key pair so must be
  generated together).

SDF Usage
---------

The `sdf` program has a flag `--cyclone-encryption-key-path` which
allows a user to specify the path to the encryption key. The default for
this flag is `/run/sdf/cyclone_encryption.key` which is where we would
place the key when running the service in a container and/or in
production.

There is an exception when running `sdf` with `cargo run` (which is how
a developer typically runs the service in development), which is the
default path is overridden to point to a pre-generated developer key
which is committed in the source tree at
`lib/cyclone/src/dev.encryption.key`. In this way, a developer doesn't
have to generate this key pair, and can start using it immediately.

Veritech/Cyclone Usage
----------------------

While the `cyclone` program ultimately uses the decryption key, the
calling of `cyclone` is configured via the `veritech` program which
manages a pool of Cyclone server instances.

Currently, `veritech` configures Cyclone instances using defaults but
this may change in the future when configuration files are introduced.
This means that `veritech` will invoke `cyclone` by setting Cyclone's
`--decryption-key` option to `/run/cyclone/decryption.key` which is
where we would deploy the decryption key when running
Veritech/Cyclone/lang-js in a container.

As with SDF, there is an exception when running `veritech` with `cargo
run`, which is that the default path is overridden to point to a
pre-generated developer key which is committed in the source tree at
`lib/cyclone/src/dev.decryption.key`. Again, both the encryption and
decryption keys are a key pair, so re-encrypted secrets in a
developer-running `sdf` service can be decrypted in a `cyclone`
instance, managed by a developer-running `veritech` service.

Impact on Function Execution
----------------------------

Now that we are managing secrets down to the function execution layer of
our stack, necessarily an encryption key is now required right alongside
a Veritch client when preparing and dispatching a function. It is this
change that has resulted in a fair amount of churn in the tests most
specifically, but...them's the breaks ;)

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>