feat(dal): Implement From trait for deculture::AttributeResolverContext & deculture::AttributeResolverContextBuilder
impl From<crate::attribute_resolver_context::AttributeResolverContext>
for AttributeResolverContext will let us easily upgrade an
AttributeResolverContext into the newer version when hooking into
existing code.

impl From<AttributeResolverContext> for AttributeResolverContextBuilder
is handy for if we have an AttributeResolverContext and we need to
"augment" it in some way, allowing us to create a builder based on it.