feat(web): hide/show headers in attribute viewer
I had a great breakdown and explanation of the implementation only to
lose the commit message contents, so now I'm dejected and simply want to
ship this change. Long story short, there's a computed `treeOpenState`
datastructure that is passed down the tree as a prop to provide state,
and a `toggleHeader` event that is emitted up from a `HeaderWidget`
which ultimately updates the `treeOpenState` so everything is reactive
accoringly.

Consult fnichol for more exciting details :)

References: hide show the headers in the attribute editor [sc-2297]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>