feat(dal): Create SQL helpers for dealing with AttributeResolverContext
This "promotes" deculture::AttributeResolverContext to be handled in a
similar way to how we're handling Tenancy, and Visibility with helpers
to deal with the SQL being given the JSON representation.

This ensures that we are _always_ checking is X "in this context" the
same way, everywhere, as well as is X "exactly for this context".