feat(dal): Add type definition to deculture::InputSocket
The `type_definition` is expected to be a JSON Schema representation
of the shape of the data that can be hooked up to an `InputSocket`
from a corresponding `OutputSocket`.  In the case of an `InputSocket`
that is created for a `Prop` (to be used internally within a single
`Component`), the `type_definition` will not be set, as the shape is
determined by the underlying `Prop`, and its children, and it is not
possible to connect the `InputSocket` to something that does not match
its shape.