feat(dal): Set parent AttributeValue when setting parent Prop for a Prop
We should only ever be setting the parent of a Prop when building a
"free-floating" tree of Props, or when building out the Props for a
SchemaVariant.

In both of these cases, we should only ever have one AttributeValue
per Prop at the time we are calling `prop.set_parent_prop`. Even in
the case where we want an Array or Map Prop to have multiple values
"by default", this won't be happening until *after* the Prop structure
has been created, while all the Props still have a value of `unset`.