chore(dal): Remove outdated Prop relationship comment
`Prop`s are going to continue to be a "belongs to" relationship to
other `Prop`s. The reason we were considering changing this to be a
many-to-many was to support including "free floating" `Prop`s into
multiple `SchemaVariant`s, where they would need to have multiple
parents.

Instead of directly including the "free floating" `Prop`s into
multiple `SchemaVariant`s, instead we are going to treat the "free
floating" `Prop`s as a blueprint, and create copies of them under the
`SchemaVariant`. This avoids potentially silently breaking usage of a
`SchemaVariant` by modifying the structure of the included `Prop`s.