feat(dal,sdf): Convert tenancies, refactor some DAL leafs
The conversion between ReadTenancy, WriteTenancy and Tenancy will help
us move from a era where only Tenancy existed to one where only
(Read/Write)Tenancy are used.

We refactored a few DAL modules functions, the conversion was used to
preserve the boundary. Originally we thought about receiving
`impl Into<(Read/Write)Tenancy>`, but Read needs txn and is await,
and Write gets a little awkward with cloning ergonomy, so we decided to
be explicit on the conversions.

The biggest refactor will be on the standard model.