feat(dal/web): Split schematic panels + better reactivity
Properly update schematic panel when visibility changes, edit session is written to. Clears visibility when application is left.

Fix schematic panel select dropdown.

Disable add menu if in component schematic panel and no deployment node is selected. Filter component nodes to only show those that implement the selected deployment node (contain an edge with its implementation input socket), no deployment node selected means the panel should empty and adding nodes is disabled (for now node creation will not connect the component to the selected deployment, so we aren't enforcing this as we should, but the next PR will handle component creation).

After clicking the node it will be kept selected, and it will sync the other panel (if using the same schematic subpanel). Deselecting also is synced. A `deploymentSelection$` and a `componentSelection$` observers were created to keep them in sync. And also to help the other panels to give relevant information about the nodes in question.

Fix fake node id (mock node used during creation dragging), it's now -1, so we filter with it.

Inherit SchematicKind/NodeKind from SchemaKind.