feat(dal): Update AttributeValue::insert_for_context to work with Arrays
If a key is not explicitly provided for an Array element, it will
automatically be given a UUID v4 value as its key. By ensuring that
Array elements have a key, we are able to use the exact same logic for
querying to find the set of values for a Prop in a given context as we
use for finding Hash elements, as the uniqueness is done based on the
PropId, and the key of the value.