feat(dal/sdf/web): Attribute panel node selection and locking
Sync attribute panel's selected component with the node's selection in the UI.

Allow locking the attribute panel  component selection through the UI, provide dropdown to
change component while locked in a controlled way.

Add a new NodeKindWithBaggage that allows NodeView to inform the frontend of
whichever component it's tied to (if any).