feat(web,dal): Show deployment node in component panel
Node positions are now always handled as a vector with two contexts,
the first context already was there and was related to systems and
applications. The second context now determines which panel the position
refers to.

Component schematic panels need to show the deployment node that is it's
parent, and the parent node needs its own position, unrelated to its position
in the deployment panel. But also eventually we will be able to link components
in multiple places, to allow reuse. So some node may appear in two component
panels, filtering by whichever deployment node is the parent of the panel.

The front-end now needs to find the appropriate position from the node positions
vec instead of getting the first one available.

This meant fixing a bunch of hacks we had through the code as we weren't using
those features, but the code quality is still lacking. We have been iterating
over this code heavily adding features without major refactors, so the quality
is gradually decreasing, it might require a _full_ refactor eventually, but I
don't think now is the time.

This PR exacerbated a bug that existed before, sometimes the selection border
doesn't get propagated to the other panel when the panel kind is being changed,
the bug will be fixed in the next bug fixing iteration as it will require
messing with code that is still _under construction_ (like the locking scape-
hatch)