feat(web): Move from local to global selection storage
We had problems syncing multiple component panels that were locked
to different nodes. And had a lot of edge-cases while syncing both
panels when they were in weird states, as the previous architecture
was limited, because it would only store selection data of the local
panel. But when we change the panel kind, or the panel deployment parent
we had to find to appropriate state for that selection, or reset it,
losing some state that was useful because we couldn't store it.

Now SelectionManager holds a SelectedNode[] instead of a Node[],
the selected node has meta-data defining which kind of panel it is.
If the parentDeploymentNodeId is null, it's a deployment panel,
otherwise the parentDeploymentNodeId will represent to which component
schematic panel the selection data belongs to.

We still assume only one selection per panel, although the code to support
more is there, we are finding the first selection of the panel and using it.

This refactor fixed a lot of edge cases, although it's still fragile, as
its fresh and needs some battle testing to find other edge cases, but the
idea is that the edge cases will be much easier to solve with this new
structure, as we always have the data we need, we just need to refresh the screen.

The code quality is still lacking, it's better now but it's not 100%. But it's
not clear if now is the time to "fix" it. I will think about it while doing the
next task (locking scape-hatch with the select box in the component schematic panel).

Also fixed a bug in sdf services leaking observables, changing shareReplay to care about
refCount fixed it.