feat(web): Add selection box scape to component panel lock
Added a select-box to PanelSchematic to choose the component.

We obtain the selected component id from the select box in
PanelSchematic and pass it to SchematicViewer, which then
uses its schematicData to convert the componentId to a nodeId.

We pass that nodeId to Viewer bypassing the lock as we hold it,
determining the selectedDeploymentNodeId.

PanelSchematic also monitors the deployment selection to sync
the select box whenever unlocked.

We don't sync de-selection with AttributePanel's select-box.

We also reset some forgotten state on visibility changes.

Cleaned up some old vue style while defining props.