fix(web): Fix schematic panel bugs
- Fix node Add button alignment.
- Change UiMenu structure for component schemas.
- Render connections below nodes (and its sockets).
- Render sockets in the node's top/bottom for the deployment schematic panel.

Notes:

I wasn't able to make a UiMenu with more than 2 layers as it expects a schema to represent a layer and it doesn't make sense to create a docker schema just to put it there, I feel I need to invest more time to learn it and fix it if it's really a bug. Instead of just lack of knowhow (I think Adam is the one that mostly knows how UiMenu works). It's still way better than before (and closed to the demo).

Socket position is still weird, it was hackish I haven't changed much, just expanded the hack to support vertical sockets. They might not work well with arbitrary socket numbers.