feat: add support for multiple systems (and no system too!)
This change adds support for creating one or more systems in the
application editor and adds the ability to work with objects that are
not in a system. While not all object management features have yet been
written (in other words, we can't move objects created outside a system
into a system later), this work provides the foundation for such
functionality to be unlocked.

A system is no longer created by default with a new billing account is
signed up for so every billing account starts fresh with zero systems
associated. This also includes the DAL test suite which was updated to
remove any such assumptions in order to shoot for correctness ;)

As most other objects in the SI domain, a system is created via a change
set/edit session, and therefore the system selector will not display the
`-- new --` option unless the user is in an edit session. The modal
dialog and behavior is similar to creating a new application including
enter key support, cancelling etc.

In the front end code, we have already established a `system$`
observable which tracks the currently selected system and is there a
`System | null` type. Note that we are tracking a full system object and
**not** a system id. In this way, the front end code knows when we have
a legitimate system active and when we do not. This allows the front end
to either omit `system_id` fields in certain SDF requests and/or provide
a `-1` value where appropriate. (NOTE: it turns out that with most of
the current front end API calls, it is easier to omit a system id than
crafting an explicit "none" value which is somehow comforting in that
most of the relevant SDF handlers in this situation already take a
`system_id: Option<SystemId>`).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>