feat(macros,sdf,dal,web): application NodeId in DalContext
Filter `Schematic::find` and `list_components_identification`'s
SDF service by application.

Add `application_node_id: Option<NodeId>` to the `DalContext`.

Always send application node id from front-end to sdf as the
ApplicationId header.

Frontend doesn't pass `rootNodeId` as JSON anymore.

Dal tests DalContext now come with an application.

If there isn't a system (-1 id), we ignore the resource and the
node->system edge creation when the node is created. This is
specially useful for applications, as they must be orthogonal
to systems, and need no resource syncing.

If SDF tests find a rootNodeId property in the JSON payload,
they use it to generate a ApplicationId header.

Improve some tests to ensure data isn't leaked across applications.

The biggest edge case I see is with SDF tests, where we can't
transparently pass the `applicationNodeId`, nor the `workspaceId`,
so we need to have fields in the JSON to pass them through, and a
hack in the test harness to convert these fields into headers.
So the field is never read by SDF, but it needs to be defined in
the SDF handler's request structure.