feat: use a signup secret when creating billing account
This change migrates the web app away from an HTTP basic authentication
password to a shared secret provided in the form during signup.

In SDF, there is a new configuration parameter `signup_secret` which is
the string which must be provided when signing up for a new billing
account. By default (i.e. for development and testing convenience) the
shared secret is `cool-steam`. This value can be overidden in SDF's
config file or with the `SI_SDF__SIGNUP_SECRET` environment variable.

References: ENG-99

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>