feat(web): Make grid infinite
Rethink vector shader so instead of a 800x800 grid it renders a
square grid with the biggest renderer side * 2 (to deal with zoom).
Instead of moving it around with the camera, it modulos the translation,
so it only moves by 10 squares, and then resets the panel to the
original position. So 1 sized grid just emulates a infinite grid.

Update frag shader so it takes zoom into account when setting
the individual square size. The zoom worked on the previous
version, but the refactor exposed the bug for some reason.