feat(web,dal): Improve code generation UI
Add copy button, to send code generated to the copy paste buffer.
This isn't the most standardized behavior ever, so there are
questions about how it works across all platforms.

Make code read-only, as the backend doesn't take user input on it
yet.

Force code generation to run when component is created (no system)
and when component gets a system, so the code isn't empty when it
shouldn't.

YAML.stringify has weird behaviors for empty object, so we fix that.