feat(web,dal): Inform user that code gen is executing
Uses an empty fbrv to signify that the code generation is being updated.
So the user doesn't depend on stale values, and effectively see it's "Loading".

For now we do not handle failures, the user will see loading forever,
they will have to change some field (and change it back) to trigger a re-run.

The next step is improving failure recovery.