bug(app): roll back codemirror and fix refresh
Rolls back codemirror to 0.19, until there is a release of the streaming
language backend to match the codemirror 0.20 releases from last week.

This also fixes the way the refresh icons work - they were animating the
rotation of the button, rather than the icon, which caused the icon to
appear as if it was spinning off its axis a little. It also now tracks
the animation state to the length of the refresh - so it stops spinning
when the call completes. It also now uses the tailwind animation
classes, rather than hand-rolled javascript applied to the html element
through a reference.