feat(app/web): fix signup flow, convert to tailwind ui
This makes the sign up/sign in more clear. It uses links to let you
navigate between signing up and signing in, and provides helpful text for
each field.

It also converts most of the UI to be more "tailwind"-esque. The form
styles, select boxes, etc. have been replaced with versions from
tailwind, with minimal modification (pretty much only things like size).
This is a first step to getting the entire UI to be built using a
reasonable design system, making it easier to tweak and change things
more globally.

Lots more whitespace, more accessible, chonkier. It looses a bit of the
edge that the old UI had, but gains a lot of "correct" behavior.

There are two known issues:

* The log in form navigates away when there is an error. I do not know
  why.
* There are still not validations on the sign up and login pages. We
  should add them to the backend, and then duplicate them on the
  frontend. That's for another day.

Love,
Adam