feat(web): Some attribute sub panels cant be reactive to visibility
Their parents actually are responsible for re-mounting them when
visibility changes, as they depend on Component selection and with the
visibility change the selection might not exist, but they don't have
enough information to find-out. Their parents do.

So the parent monitors for visibility changes, if the selected component
is gone it de-selects, and only then retriggers queries like
getEditFields and listQualifications in the sub-panels, by either
updating the selected component, or changing its key.

Fixes the old implementation that ignored edit session changes, so it
wasn't properly refreshing the fields when it should, now it does.

I commented out the reactivity logic in the sub-panels as I understand
this isn't a great approach, but it works and it's there, if we ever
decide that it's a bad one we can easily revert.