feat(web): Ignore invalid visibility in some queries
Before when visibility changed and the selected component wasn't visible
anymore the user would get an error, but it's an acceptable state, we
just have to ignore the failure.

With that in mind we changed getEditFields and listQualifications to
ignore when the tenancy is correct but the visibility is not, as the
cache might be stale.

Also removed the old reactivity hack we had to solve this in the
frontend instead of the backend