feat(web): add validation and form blocking
Adds an SiValidation atom that takes a prop of validation functions to
run. Used in combination with an SiTextBox2, it creates an easy way to
validate forms without presupposing too much about layout or style.

The validation widget understands that you shouldn't see errors if the
field has not been "dirtied" (typically by simply blur-ing it), shows
`(required)` in the label if the field is required, and takes arbitrary
validation functions.

Also adds a "composables" directory, to hold vue 3 composable functions.
The useFieldErrors composable is related to keeping track of the errors
states across forms, which makes it easy to then enable/disable the
submit button if there are errors.

Adds [validator.js](https://github.com/validatorjs/validator.js) as a
dependency, which gives us access to a large number of pre-built
validation functions.