chore(*): typescript types for property editor
Adds basic types for a property editor that is still generic, but
focused on getting the correct types returned rather than *how* the
types are built.

It also fixes an issue with how we configure what URL to hit for backend
services; it's now based on the URI's host field if you aren't in
production.

Also adds a basic property editor that is "flat".