feature(web): basic property stub editor
This PR has a basic stub of a property editor, that implements a "flat"
editing structure. It uses labels on headers to denote nesting of
elements, rather than trying to nest the property editor itself using
whitespace.

It is not hooked up to the backend yet - but it does have examples of
updating values, adding to maps and arrays, and having complex
properties. It's ready to be hooked up to real data.