feat(web): add booleans/checkbox widgets to property editor
This change adds an `SiCheckBox` atom that is similar is scope and
behavior to the `SiTextBox2` for text fields and a primary consuming
component `WidgetCheckBox` to handle the remaining widget behavior.

Note that unlike prior implementations, this checkbox version can accept
a raw value that is `null`, leading to a non-checked (i.e.
false-by-default) checkbox that is set into an [indeterminate state].
While this is primarily a display-only concern (as far as HTML/DOM
goes), it means that we can signal when an unset value is present for a
boolean. Additionally this means that we can set the checkbox state to
checked (i.e. set `true`) or unchecked (i.e. set `false`) and then use
the unset button to set the underlying value to `null`.

[indeterminate state]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#indeterminate_state_checkboxes

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>