bugchore(dal,sdf): fix Prop.set_parent_prop behavior; delete edit fields; fix root prop for schema creation
Sorry for conflation of fixes, but they all ended up being somewhat
related:

`set_parent_prop()`
-------------------

Previously, we were creating values for the context that this method was
provided and creating a relationship between them. Now, we are only
dealing with values at the prop level of specificity when creating the
relationship.

Deleting Edit Fields
--------------------

Guess what, nobody uses these anymore! Also when correcting the above
behavior we were failing tests that were using edit fields to verify
other non-edit-field behavior. Note that somewhere in the bowls of the
frontend, we still have a "schema editor" which will no longer function,
but there is currently no navigation to it so...no one's using it
anyway.

Fix Root Prop for Schema Creation
---------------------------------

When setting the base values for the set of root props on a schema, do
so at the prop level of specificity.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>