feat(web): filter container child value types from api
This change removes child entries in
`propertyEditorContext.values.childValues` whose entries are *not*
elements of the parent array or map values of the parent map.

Ideally the backend api should filter these out, but until then the
front end will by looking for children of arrays and maps that have a
`null` key (meaning that they are not element/value members of the
collection--neat trick, no?).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>