bug(dal): Ensure ComponentView processes the work queue front-to-back
Previously, we were always popping off the back of the work
queue (treating it as a stack), even though we had ordered it to be
processed front-to-back.

By moving the list of items to be processed, after we sort it
appropriately, into a VecDeque, we are able to pop work items off of
the front, and treat the work queue as a proper queue.