feat(web,dal): Dim nodes that can't be connected to
This entailed in a major cleanup and rewrite of a bunch of front-end parts. Not everything was touched, but the frontend is way cleaner now.

Since the backend isn't there yet we are currently hacking away the provider matching data, merging local metadata with the current dynamic schematic.

This means we have some hacks, as for example we can't know the schema information if a component for that schema hasn't been created, so the first node of a schema will appear without sockets during creation. After creation we will have it and it will be instantly updated.

We also have some hacks to deal with schemas that aren't available. And probably are hiding some sockets that appeared before (like the service ones, but since they can't connect to anything they weren't really useful).