chore(web): Refactor selection logic
Significantly cleanup selection logic, it used to be a hybrid of global
state, so panels could share selection, and local state, so panels could
tie a specific rendered node to a selection.

This turned out to create a very complex logic, so we completely
stripped the local concept of selection, now when the global state
changes we just walk through the children nodes of a panel, updating
selection state for each node. Nothing is stored locally.

Minor misc fixes:
- Swap locked/unlocked icon and fix color in the locked one
- Update SiSelect value type to supress a type error
- Turn a few variables into non null
- Viewer doesn't need to know if panel is locked, now graphical node
selection state is completely separate of parent deployment node selection